#146 Add SwiftPM plugin for `UnitIdentifier` generated code
