+++ Sources/Bluetooth/Resources/CompanyIdentifier.json
