+++ Sources/GenerateBluetooth/Generate.swift
@@ -44,6 +44,33 @@ struct GenerateTool {
                 input: inputFile,
                 output: outputFile
             )
+        case .unitIdentifier:
+            // parse arguments
+            guard arguments.count == 5 else {
+                throw CommandError.invalidArguments(arguments)
+            }
+            let inputFile = URL(fileURLWithPath: arguments[2])
+            let outputFiles = [
+                URL(fileURLWithPath: arguments[3]),
+                URL(fileURLWithPath: arguments[4])
+            ]
+            // generate files
+            try generateUnitIdentifiers(
+                input: inputFile,
+                output: outputFiles
+            )
+        case .unitIdentifierTests:
+            // parse arguments
+            guard arguments.count == 4 else {
+                throw CommandError.invalidArguments(arguments)
+            }
+            let inputFile = URL(fileURLWithPath: arguments[2])
+            let outputFile = URL(fileURLWithPath: arguments[3])
+            // generate files
+            try generateUnitIdentifierTests(
+                input: inputFile,
+                output: outputFile
+            )
         }
     }
 }
@@ -52,6 +79,8 @@ enum CommandType: String {
     
     case companyIdentifier
     case companyIdentifierTests
+    case unitIdentifier
+    case unitIdentifierTests
 }
 
 enum CommandError: Error {
